Direct Message ids are now long

+++ twitter4j-core/src/main/java/twitter4j/DirectMessageJSONImpl.java
@@ -38,7 +38,7 @@ import static twitter4j.ParseUtil.*;
  * @author Yusuke Yamamoto - yusuke at mac.com
  */
 /*package*/ final class DirectMessageJSONImpl extends TwitterResponseImpl implements DirectMessage, java.io.Serializable {
-    private int id;
+    private long id;
     private String text;
     private int senderId;
     private int recipientId;
@@ -56,7 +56,7 @@ import static twitter4j.ParseUtil.*;
         init(json);
     }
     private void init(JSONObject json) throws TwitterException{
-        id = getInt("id", json);
+        id = getLong("id", json);
         text = getUnescapedString("text", json);
         senderId = getInt("sender_id", json);
         recipientId = getInt("recipient_id", json);
@@ -74,7 +74,7 @@ import static twitter4j.ParseUtil.*;
     /**
      * {@inheritDoc}
      */
-    public int getId() {
+    public long getId() {
         return id;
     }
 
@@ -156,7 +156,7 @@ import static twitter4j.ParseUtil.*;
 
     @Override
     public int hashCode() {
-        return id;
+        return (int)id;
     }
 
     @Override
